Add files via upload
Here we model the daily DEenergy load using SARIMA. The goal is to fit a model that captures both the weekly seasonal pattern e.g., there is a significant drop in energy usage mostly on Sundays.  Our model seems to fits the data only partially is not fully able to capture the heteroskedastic residual variance.