Add files via upload
Here, we model the daily DEenergy load using SARIMA. The goal is to fit a model that captures the weekly seasonal pattern as there is a significant drop in energy usage mostly on Sundays.  Our model seems to fits the data only partially and it is not fully able to capture the heteroskedastic residual variance.