Add files via upload
This file contains a serie daily dataa computed as the average of hourly data. 
The variables found in the file are self explainatory, apart from the temperature_price_interaction, which was obtain by multiplying price data point with temperature data point for each day.
These features where used used to model the SARIMAX 2, the State Space Kalman Filter , the Gaussian Process, and the Time Foundation.