Add files via upload
The SARIMAX was use in our project as a baseline model. Note that although the model fail to reject Heteroskedascity, we should be careful in not assuming directly volatility clustering and applying models like the SARIMAX-GARCH because this would cause misspecification. Instead, we used more advanced and machine learning models; some of which proven to be very fitting.
Please check the power point presentation for key notes.
By K.Fane