Add files via upload
The SARIMAX was use in our project as a baseline model. Note that in comparison to SARIMAX_2 weekly load forecast with the average hourly data to obtain daily data, the present weekly load forecast using directly the hourly data is performing slightly better in terms of no heteroskedascity, although still not significantly. Again, we should interpret this result as a proof of volatility clustering and apply models like the SARIMAX-GARCH because this would cause misspecification. Instead, we used more advanced architectures for our forecast.
We also note that in terms of MPAE, the previous model presents a smaller value but also uses 16% more data points. 
Please check the model, and the power point slides for more.
By K.Fane