Add files via upload
In this model, we forecast the load using Kalman Filter. What is unique about it is that we used expansion window in the testing stage of the model. This rendered the model a dynamic learner and provided high fitted results.
Please check the model and the power point presentation slides.
By K.Fane