Add files via upload
An other advanced architecture used to forecast energy load is the the Time foundations model. Just like the Gaussian process, it uses transformers at its core. There is room for the improvement of this model as there seems to be some seasonal pattern visible in the residual subplot. We will improve the model ulteriorly. Please see the model and the power point presentation slides.
By. K.Fane