Add files via upload
In this model, we forecast the load using Kalman Filter. What is unique about it is that we used expansion window while forecasting load of the testing subset. This makes the model learn dinamically as more data is fed in.
Please check the model and the power point presentation slides.
By K.Fane